+++ src/services/salesAd/deleteSalesAd.service.ts
@@ -2,6 +2,7 @@ import { salesAdRepository } from "../../data-source";
 import { Ad } from "../../entities/ads.entity";
 
 const deleteSalesAdService = async (id: number): Promise<void> => {
+
   const salesAd: Ad | null = await salesAdRepository.findOne({
     // @ts-ignore
     where: {
